+++ packages/blocks-core/CHANGELOG.md
@@ -28,6 +28,7 @@
 - `Columns` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Accordion Heading` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Comment Content` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
+- `Post Excerpt` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 
 
 ### Automated Tests
@@ -57,6 +58,7 @@
 - Added E2E tests to check functionality `Columns` new inner block.
 - Added E2E tests to check functionality `Accordion Heading` new inner block.
 - Added E2E tests to check functionality `Comment Content` new inner block.
+- Added E2E tests to check functionality `Post Excerpt` new inner block.
 
 ## 1.4.0 (2025-05-19)